+++ utils/kafka-test-utils/src/main/scala/pl/touk/nussknacker/test/containers/kafka/WithSchemaRegistryContainer.scala
@@ -1,6 +1,5 @@
 package pl.touk.nussknacker.test.containers.kafka
 
-import com.dimafeng.testcontainers.GenericContainer
 import org.scalatest.Suite
 import pl.touk.nussknacker.test.containers.WithDockerContainers
 
@@ -8,22 +7,13 @@ import java.util.Arrays.asList
 
 trait WithSchemaRegistryContainer { self: Suite with WithDockerContainers =>
 
-  private val schemaRegistryPort         = 8081
   private val schemaRegistryNetworkAlias = "schema-registry"
 
-  protected val schemaRegistryContainer: GenericContainer =
-    new GenericContainer(
-      "ghcr.io/axonops/axonops-schema-registry:0.2.1",
-      exposedPorts = Seq(schemaRegistryPort),
-    ).configure { self =>
+  protected val schemaRegistryContainer: SchemaRegistryContainer =
+    SchemaRegistryContainer().configure { self =>
       self.setNetwork(network)
       self.setNetworkAliases(asList(schemaRegistryNetworkAlias))
-      self.withLogConsumer(logConsumer(prefix = "schema-registry"))
+      self.withLogConsumer(logConsumer(prefix = schemaRegistryNetworkAlias))
     }
-
-  // testcontainers exposes schema registry via mapped port on host network, it will be used for kafkaClient in tests, signal sending etc.
-  protected def hostSchemaRegistryUrl: String =
-    s"http://${schemaRegistryContainer.host}:${schemaRegistryContainer.mappedPort(schemaRegistryPort)}"
-  protected def containerSchemaRegistryUrl: String = s"http://$schemaRegistryNetworkAlias:$schemaRegistryPort"
 
 }