+++ docs/Changelog.md
@@ -362,6 +362,8 @@ description: Stay informed with detailed changelogs covering new features, impro
 * [#9286](https://github.com/TouK/nussknacker/pull/9286) Fixed handling of Avro records in Flink SQL component.
 * [#9322](https://github.com/TouK/nussknacker/pull/9322) Added mockable Azure Schema Registry client
 * [#9327](https://github.com/TouK/nussknacker/pull/9327) Remove bind mounts from Docker tests
+* [#9336](https://github.com/TouK/nussknacker/pull/9336) Moved traits for starting Kafka and Schema Registry containers
+  from Flink-specific `nussknacker-flink-test-utils` to `nussknacker-kafka-test-utils`
 
 ## 1.18